+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - 添加用户信息管理测试用例 | Add user information management test cases
 - 添加更改密码测试用例 | Add change password test case
 - 添加资源文件测试用例 | Add asset file test case
+- 实现草稿更新功能 | Implement draft update feature
 
 ## 2024-04-01